backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic[email]==2.10.0
boto3==1.35.0
python-multipart==0.0.12
anthropic==0.40.0
firebase-admin==6.7.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-dotenv==1.0.0
httpx==0.27.0
pytest==8.0.0
pytest-asyncio==0.23.5
hypothesis==6.98.0
